feat: analyze-image skelton
